refactor introcut
